Add fundamental sharding types
