+++ Cargo.toml
@@ -92,6 +92,7 @@ jsonwebtoken = "8"
 libc = "0.2"
 md5 = "0.7.0"
 memoffset = "0.8"
+mur3 = "0.1.0" 
 native-tls = "0.2"
 nix = "0.26"
 notify = "5.0.0"
