+++ libs/pageserver_api/Cargo.toml
@@ -17,5 +17,7 @@ postgres_ffi.workspace = true
 enum-map.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
+mur3.workspace = true
+url.workspace = true
 
 workspace_hack.workspace = true
